# Improve `TrainingDataset __getitem__` performance