--- a/src/boltz/data/tokenize/boltz.py
+++ b/src/boltz/data/tokenize/boltz.py
@@ -1,4 +1,4 @@
-from dataclasses import astuple, dataclass
+from dataclasses import dataclass
 
 import numpy as np