--- a/src/boltz/data/tokenize/boltz2.py
+++ b/src/boltz/data/tokenize/boltz2.py
@@ -249,8 +249,8 @@ def tokenize_structure(  # noqa: C901, PLR0915
                 token_data.append(token_astuple(token))
 
                 # Update atom_idx to token_idx
-                for atom_idx in range(atom_start, atom_end):
-                    atom_to_token[atom_idx] = token_idx
+                atom_to_token.update(
+                    dict.fromkeys(range(atom_start, atom_end), token_idx))
 
                 token_idx += 1
 
@@ -348,8 +348,8 @@ def tokenize_structure(  # noqa: C901, PLR0915
                 token_data.append(token_astuple(token))
 
                 # Update atom_idx to token_idx
-                for atom_idx in range(atom_start, atom_end):
-                    atom_to_token[atom_idx] = token_idx
+                atom_to_token.update(
+                    dict.fromkeys(range(atom_start, atom_end), token_idx))
 
                 token_idx += 1