@jwohlwend 2025-07-10
Cool stuff, thank you!